Supervise all Recorders and PlayBacks
